controller: add lib to transform from image to world space
The lane detection computer vision model outputs pairs of
x, y coordinates in image space.
Different methods, such as the MPC, requires these pairs of
points to be in world space.
It's possible to transfrom from image to world space using the
camera's intrinsic and extrinsic parameters and assuming that
the points lie on a flat horizontal plane.